jv/requirements.txt
flask==2.3.3
pywebview==4.2.2
